refactor: Move SearchModule to modules dir
